feat:1/2

+++ src/main/java/org/seec/muggle/auror/po/Role.java
@@ -5,8 +5,9 @@ import lombok.Data;
 import java.util.List;
 
 @Data
-public class Role {
-    private Integer id;
+public class
+Role {
+    private Long id;
     private String name;
     private List<User> users;
     private List<Permission> permissions;
